requirements.txt
google-genai>=1.0.0
supabase>=2.0.0
clerk-sdk-python>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
python-multipart>=0.0.12
httpx>=0.28.0
Pillow>=11.0.0
# MusicGen BGM 생성용
transformers>=4.40.0
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.10.0
